refactor(incidents): extract InsightCard and badges to shared module
